Add GCP and Azure support to agent-pool-terraform
Closes #34040

The intention of this change is to keep the overall structure of the
guide and Terraform module close to the original in order to minimize
the scope of the change and reuse text and Terraform configurations
between the different cloud providers.

There isn't an elegant way to dynamically enable and disable `provider`
and `required_providers` settings based on input variables, so the
choice was between (a) writing separate modules for each cloud provider
and (b) having the user modify the module to remove unneeded cloud
provider configurations. I opted for (b) to allow the module to reuse
resources and the cloud-init script.

While adding GCP support was straightforward, Azure was less so since
the Azure VM resource requires an SSH public key. I made the SSH public
key file path an input variable while configuring the cloud-init script
disable OpensSSH and delete any authorized public keys on the instance.

This change also improves the module and guide in two ways that are
unrelated to adding Azure and GCP support:

- Adds an AWS AMI data source instead of hardcoding an AMI ID for EC2
  instances. The hardcoded AMI has already become unavailable.
- Clarifies the requirement for NAT infrastructure so instances can
  communicate with the Teleport Proxy Service, and explains that demo
  environments can assign public IPs to each instance if NAT
  infrastructure is unavailable.